small adjustments so it runs on GNU compiles on home setup

+++ commander1/src/commander/comm_system_backend.cpp
@@ -45,7 +45,7 @@ extern "C"
 	{
 		stringstream path; path << "/proc/" << getpid() << "/statm";
 		FILE * f = fopen(path.str().c_str(),"r");
-		int64_t npage;
+		long long int npage;
 		fscanf(f, "%lld", &npage);
 		fclose(f);
 		return npage*getpagesize();
